move train and inference wrappers to separate module
